fix: KEEP-219 move CCIP receiver transform to encode-transforms module
The workflow bundler tree-shakes side-effect imports from "use step"
files. Transforms registered in chainlink.ts (via import "@/protocols")
were not surviving the production build, causing the receiver address
to be encoded as 20 raw bytes instead of 32 abi-encoded bytes.

Moving the registration to protocol-encode-transforms.ts ensures it
executes because the step handlers import applyEncodeTransformsNamed
as a named import, which the bundler preserves.
